--- src/components/ProgressBar/components/Percentage.tsx
@@ -1,24 +1,0 @@
-import React from "react";
-import classNames from "classnames";
-
-type PercentageProps = {
-  percent: number;
-  percentageAfter: boolean;
-  customText?: string;
-};
-
-export const Percentage = ({ percent, percentageAfter, customText }: PercentageProps) => {
-  const percentageClassNames = classNames({
-    "percentage-after": percentageAfter,
-  });
-
-  return (
-    <div
-      data-testid="percentage"
-      className={percentageClassNames}
-      dangerouslySetInnerHTML={{
-        __html: customText || `${percent}&rlm;%`,
-      }}
-    />
-  );
-};
