+++ src/components/ProgressBar/types.ts
@@ -1,0 +1,2 @@
+export type Size = "md" | "lg" | "sm" | "xs" | number;
+export type Color = "primary" | "success" | "white" | "darkgreen";
